# openapi/brindle.v1.yaml
openapi: 3.1.0
info:
  title: Brindle
  version: '2026-06-15'
  summary: Transactional email and SMS over one HTTP API.
  description: |
    Brindle takes a message, renders it, hands it to a carrier and tells you what happened to it.

    Every request under `/v1` carries `Authorization: Bearer <key>`. Keys begin `bk_live_` or
    `bk_test_`; a test key is accepted and everything it does is recorded, and nothing leaves
    Brindle. Every answer carries `Brindle-Version`, which is the dated version of the API this
    deployment speaks, and the `X-RateLimit-*` headers for the account the key belongs to.

    Refusals are the same object everywhere: `type`, `title`, `status`, `detail` and `code`. Branch
    on `code`; `detail` is written for a person and may be reworded.
  contact:
    name: Brindle support
    email: support@brindle.dev
    url: https://docs.brindle.dev
servers:
  - url: https://api.brindle.dev
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/messages:
    post:
      summary: Send a message
      operationId: sendMessage
      tags: [Messages]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendRequest'
            examples:
              fromATemplate:
                summary: From a template
                value:
                  to: esme.ingham@brookmail.com
                  channel: email
                  template_id: tmpl_13jprr39jzhs
                  variables:
                    first_name: Esme
                    order_number: W-4410
                  tags: [delivery]
              writtenOut:
                summary: Written out in the request
                value:
                  to: '+447700900123'
                  channel: sms
                  body: Your bike is ready to collect until Saturday.
      responses:
        '201':
          description: The message was accepted and is queued.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      summary: List messages
      operationId: listMessages
      tags: [Messages]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/MessageStatus'
        - name: channel
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Channel'
      responses:
        '200':
          description: A page of messages, newest first.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Message'
                  next_cursor:
                    $ref: '#/components/schemas/NextCursor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/messages/batch:
    post:
      summary: Send up to five hundred messages
      operationId: sendBatch
      tags: [Messages]
      description: |
        A batch is not all or nothing. Rows that can be sent are sent; rows that cannot come back
        with their position in the array and the code that row would have been refused with on its
        own.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items:
                    $ref: '#/components/schemas/SendRequest'
            examples:
              two:
                value:
                  messages:
                    - to: colm.wardle@brookmail.com
                      channel: email
                      subject: Your order is ready
                      body: <p>Come and get it.</p>
                    - to: '+447700900771'
                      channel: sms
                      template_id: tmpl_1hjhpvx6q447
                      variables:
                        code: '4471'
      responses:
        '202':
          description: The batch was taken. Read `rejected` before you treat it as sent.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/messages/{id}:
    parameters:
      - $ref: '#/components/parameters/MessageId'
    get:
      summary: Read one message
      operationId: getMessage
      tags: [Messages]
      responses:
        '200':
          description: The message.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/messages/{id}/events:
    parameters:
      - $ref: '#/components/parameters/MessageId'
    get:
      summary: What happened to a message
      operationId: listMessageEvents
      tags: [Messages]
      description: |
        The whole history, oldest first. A message has a handful of events, so this is not a page.
      responses:
        '200':
          description: The events, oldest first.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageEvent'
              examples:
                bounced:
                  value:
                    - id: evt_7gsp79b2gn3w
                      type: queued
                      occurred_at: '2026-09-01T20:56:11.000Z'
                      detail: null
                    - id: evt_9nh1ft4swvvm
                      type: sent
                      occurred_at: '2026-09-01T20:56:13.000Z'
                      detail: null
                    - id: evt_pnf5b6n1zwnq
                      type: bounced
                      occurred_at: '2026-09-01T20:56:19.386Z'
                      detail: The receiving server said the address does not exist.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /v1/suppressions:
    get:
      summary: List suppressed addresses
      operationId: listSuppressions
      tags: [Suppressions]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of suppressions, newest first.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Suppression'
                  next_cursor:
                    $ref: '#/components/schemas/NextCursor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    post:
      summary: Suppress an address
      operationId: createSuppression
      tags: [Suppressions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [address]
              properties:
                address:
                  type: string
                  description: An email address or a phone number in international form.
                reason:
                  $ref: '#/components/schemas/SuppressionReason'
            examples:
              askedUsToStop:
                value:
                  address: zara.tewson@brookmail.com
                  reason: unsubscribed
      responses:
        '201':
          description: The address is on the list.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Suppression'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /v1/suppressions/{address}:
    parameters:
      - name: address
        in: path
        required: true
        schema:
          type: string
        description: The suppressed address, percent-encoded.
        example: zara.tewson%40brookmail.com
    delete:
      summary: Take an address off the suppression list
      operationId: deleteSuppression
      tags: [Suppressions]
      responses:
        '204':
          description: The address is off the list.
          headers:
            Brindle-Version:
              $ref: '#/components/headers/BrindleVersion'
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimitLimit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimitRemaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer bk_live_...`. A `bk_test_` key is accepted everywhere a live key is,
        and nothing it sends leaves Brindle.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: The `next_cursor` from the page before this one.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
      description: How many rows to return. Twenty-five otherwise, one hundred at most.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 255
      description: |
        A key of your own. The first request under it is run and its answer kept for a day; a
        repeat under the same key returns that answer with `Idempotency-Replayed: true`. The same
        key with a different body is refused with `idempotency_conflict`.
    MessageId:
      name: id
      in: path
      required: true
      schema:
        type: string
      example: msg_901pn19hh7xt
...
  headers:
    BrindleVersion:
      description: The dated version of the API this deployment speaks.
      schema:
        type: string
        examples: ['2026-06-15']
    IdempotencyReplayed:
      description: Present and `true` when this answer was kept from an earlier request.
      schema:
        type: string
        examples: ['true']
    RateLimitLimit:
      description: Requests this account may make in a window.
      schema:
        type: integer
    RateLimitRemaining:
      description: Requests left in the current window.
      schema:
        type: integer
    RateLimitReset:
      description: |
        The Unix second at which the current window ends. A window opens on the account's first
        request and runs for a minute.
      schema:
        type: integer
  responses:
    BadRequest:
      description: The request could not be read.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            notAnAddress:
              value:
                type: https://docs.brindle.dev/errors/invalid_request
                title: Invalid request
                status: 400
                detail: 'to: "wren at brookmail" is not an email address.'
                code: invalid_request
    Unauthorized:
      description: No key, or a key Brindle does not know.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            noKey:
              value:
                type: https://docs.brindle.dev/errors/authentication_required
                title: Authentication required
                status: 401
                detail: 'Send your key as an Authorization header: `Authorization: Bearer bk_live_...`.'
                code: authentication_required
    NotFound:
      description: There is no such object on this account.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            noTemplate:
              value:
                type: https://docs.brindle.dev/errors/template_not_found
                title: Template not found
                status: 404
                detail: There is no template tmpl_nosuchthing on this account.
                code: template_not_found
    Conflict:
      description: An idempotency key was reused over a different body.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: The request was read and cannot be carried out.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            suppressed:
              value:
                type: https://docs.brindle.dev/errors/recipient_suppressed
                title: Recipient suppressed
                status: 422
                detail: ada.wardle@brookmail.com is on your suppression list. Take it off the list if you mean to write to it again.
                code: recipient_suppressed
    PayloadTooLarge:
      description: The request body is larger than this API accepts.
      headers:
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            tooBig:
              value:
                type: https://docs.brindle.dev/errors/payload_too_large
                title: Payload too large
                status: 413
                detail: That request body is larger than this API accepts.
                code: payload_too_large
    RateLimited:
      description: The account has spent its requests for this window.
      headers:
        Retry-After:
          description: Seconds to wait.
          schema:
            type: integer
        Brindle-Version:
          $ref: '#/components/headers/BrindleVersion'
        X-RateLimit-Limit:
          $ref: '#/components/headers/RateLimitLimit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/RateLimitRemaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/RateLimitReset'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Channel:
      type: string
      enum: [email, sms]
    MessageStatus:
      type: string
      enum: [queued, sent, delivered, bounced, failed]
    EventType:
      type: string
      enum: [queued, sent, delivered, bounced, opened, clicked, complaint, failed]
    SuppressionReason:
      type: string
      enum: [hard_bounce, complaint, unsubscribed, manual]
    NextCursor:
      type: [string, 'null']
      description: Pass this back as `cursor` for the next page. Null on the last page.
    Variables:
      type: object
      additionalProperties:
        type: string
      description: A value for every `{{placeholder}}` in the template.
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
          description: The stable name of the refusal. Branch on this.
    SendRequest:
      type: object
      required: [to, channel]
      properties:
        to:
          type: string
          description: An email address, or a phone number in international form for `sms`.
        channel:
          $ref: '#/components/schemas/Channel'
        template_id:
          type: string
          description: Send the current version of this template. Not with `body`.
        variables:
          $ref: '#/components/schemas/Variables'
        subject:
          type: string
          description: Required for an email written out in the request.
        body:
          type: string
          description: The message itself, when it does not come from a template.
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 40
    Message:
      type: object
      required:
        - id
        - channel
        - to
        - template_id
        - template_version
        - variables
        - subject
        - body
        - tags
        - status
        - batch_id
        - created_at
      properties:
        id:
          type: string
          examples: [msg_901pn19hh7xt]
        channel:
          $ref: '#/components/schemas/Channel'
        to:
          type: string
        template_id:
          type: [string, 'null']
        template_version:
          type: [integer, 'null']
        variables:
          oneOf:
            - $ref: '#/components/schemas/Variables'
            - type: 'null'
        subject:
          type: [string, 'null']
        body:
          type: [string, 'null']
        tags:
          type: array
          items:
            type: string
        status:
          $ref: '#/components/schemas/MessageStatus'
        batch_id:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
    MessageEvent:
      type: object
      required: [id, type, occurred_at, detail]
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/EventType'
        occurred_at:
          type: string
          format: date-time
        detail:
          type: [string, 'null']
          description: What the carrier said, on a bounce or a failure.
    BatchResult:
      type: object
      required: [batch_id, accepted, rejected]
      properties:
        batch_id:
          type: string
          examples: [batch_qt1tspng4bt8]
        accepted:
          type: array
          items:
            type: object
            required: [index, id]
            properties:
              index:
                type: integer
                description: The row's position in the array that was sent.
              id:
                type: string
        rejected:
          type: array
          items:
            type: object
            required: [index, code, detail]
            properties:
              index:
                type: integer
              code:
                type: string
              detail:
                type: string
...
    Suppression:
      type: object
      required: [address, reason, created_at]
      properties:
        address:
          type: string
        reason:
          $ref: '#/components/schemas/SuppressionReason'
        created_at:
          type: string
          format: date-time
...